Move training data config
